fix: response with correct HTTP/1.x Message format
The correct `HTTP/1.x Message`'s headers are followed after the `start-line` which describing the requests to be implemented. After the `headers` is a `blank line` indicating all `meta-information` for the request has been sent. Then the `optional body` is followed.
Increase HTTP buffer size to 1024(keep the same with the buffer size in `ch20-01-single-threaded.md`).